refactor:#172 meal 도메인 분리
