+++ src/app/(client)/meal/components/meal-images-carousel.tsx
