+++ src/app/(client)/meal/post/components/add-image.tsx
@@ -3,12 +3,12 @@ import { ChangeEvent, useEffect, useMemo, useRef, useState } from 'react';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import { useUserStore } from '@/store/user-store';
-import { getFileId } from '@/lib/utils/file.util';
 import { cleanupBlobUrl } from '@/lib/utils/cleanup-blob-url.util';
-import { MAX_FILE_SIZE, MAX_MEAL_IMAGE_COUNT } from '@/constants/meal.constant';
 import SITE_MAP from '@/constants/site-map.constant';
 import PIC_LINE from '@/../public/icons/pic_line.svg';
 import CLOSE_LINE from '@/../public/icons/close_line.svg';
+import { MB } from '@/constants/common.constant';
+import { getFileId } from '../../utils/file.util';
 
 type AddImageProps = {
   onImagesChange?: (files: File[]) => void;
@@ -140,6 +140,9 @@ type ImagePreview = {
   imageUrl: string;
   fileId: string;
 };
+
+const MAX_FILE_SIZE = 10 * MB;
+const MAX_MEAL_IMAGE_COUNT = 3;
 
 const ALERT_MESSAGES = {
   MAX_COUNT_EXCEEDED: '3장까지 첨부 가능합니다.',
