+++ src/app/(client)/set-goal/schemas/user-physical-profile.schema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-const USER_PHYSICAL_PROFILE_SCHEMA = z.object({
+export const userPhysicalProfileSchema = z.object({
   gender: z.enum(['MAN', 'WOMAN']),
   height: z.number().positive(),
   weight: z.number().positive(),
@@ -8,5 +8,3 @@ const USER_PHYSICAL_PROFILE_SCHEMA = z.object({
   activityLevel: z.enum(['VERY_LOW', 'LOW', 'MODERATE', 'HIGH', 'VERY_HIGH']),
   purpose: z.enum(['WEIGHT_LOSS', 'WEIGHT_MAINTENANCE', 'MUSCLE_GAIN'])
 });
-
-export default USER_PHYSICAL_PROFILE_SCHEMA;
