--- src/constants/meal.constant.ts
@@ -1,4 +1,0 @@
-import { MB } from '@/constants/common.constant';
-
-export const MAX_FILE_SIZE = 10 * MB;
-export const MAX_MEAL_IMAGE_COUNT = 3;
